9vx: Linux/x86-64 works
With any luck, I've broken everyone else.

+++ src/9vx/data2s.c
@@ -36,6 +36,6 @@ main(int argc, char *argv[])
 			printf("\n");
 	}
 	printf("\n\n.p2align 2\n%s%slen:\n", prefix, argv[1]);
-	printf(".long %d\n", len);
+	printf(".long %d\n", (int)len);
 	return 0;
 }
